+++ src/routes/ConsortiumManager/settings/authorization-roles/UsersCapabilities/UsersCapabilitiesCompareItems/UsersCapabilitiesCompareItems.test.js
@@ -7,8 +7,9 @@ import {
 import userEvent from '@folio/jest-config-stripes/testing-library/user-event';
 import {
   useAuthorizationRoles,
-  useRoleCapabilities,
-  useRoleCapabilitySets, useUsers,
+  useUserCapabilities,
+  useUserCapabilitiesSets,
+  useUserRolesByUserIds
 } from '@folio/stripes-authorization-components';
 
 import {
@@ -19,14 +20,17 @@ import {
 import {ConsortiumManagerContextProviderMock} from 'helpers';
 import { COMPARE_ITEM_NAME } from '../../../users/general/PermissionSets/PermissionSetsCompare/constants';
 import {UsersCapabilitiesCompareItems} from "./UsersCapabilitiesCompareItems";
+import {useUsers} from "../../../../../../hooks/useUsers/useUsers";
 
 jest.mock('@folio/stripes-authorization-components', ()=> ({
   ...jest.requireActual('@folio/stripes-authorization-components'),
-  useRoleCapabilities: jest.fn(),
-  useRoleCapabilitySets: jest.fn(),
+  useUserCapabilities: jest.fn(),
+  useUserCapabilitiesSets: jest.fn(),
   useAuthorizationRoles: jest.fn(),
-  useUsers: jest.fn(),
+  useUserRolesByUserIds: jest.fn(),
 }))
+
+jest.mock('../../../../../../hooks/useUsers/useUsers');
 
 const selectedMemberOptions = tenants.filter((el, index)=> index < 3).map(({ name, id }) => ({ value: id, label: name }));
 
@@ -37,17 +41,23 @@ const defaultProps = {
   members: selectedMemberOptions,
 };
 
-const roles =  [{ id: '1', name: 'role-1', metadata: {updatedByUserId: '11'} }, { id: '2', name: 'role-2', metadata: {updatedByUserId: '22'} }, { id: '3', name: 'role-3', metadata: {updatedByUserId: '33'} }];
-const users = {
-  11: {
-    id: '11',
+const userRolesResponse = [{
+    roleId: '1',
+    userId: '1',
+  }
+]
+
+const roles =  [{ id: '1', name: 'role-1' }, { id: '2', name: 'role-2' }, { id: '3', name: 'role-3' }];
+const users = [
+  {
+    id: '1',
     username: 'Admin1'
   },
-  22: {
-    id: '22',
+   {
+    id: '2',
     username: 'Admin2'
   }
-}
+]
 
 const wrapper = ({ children }) => (
   <MemoryRouter>
@@ -73,15 +83,18 @@ describe('UsersCapabilitiesCompareItems', () => {
     useUsers.mockClear().mockReturnValue({
       users
     })
-    useRoleCapabilitySets.mockClear().mockReturnValue({
-        groupedRoleCapabilitySetsByType,
-        initialRoleCapabilitySetsSelectedMap: 1,
+    useUserRolesByUserIds.mockClear().mockReturnValue({
+      userRolesResponse
+    })
+    useUserCapabilitiesSets.mockClear().mockReturnValue({
+        groupedUserCapabilitySetsByType: groupedRoleCapabilitySetsByType,
+        initialUserCapabilitySetsSelectedMap: 1,
         capabilitySetsTotalCount: 2
       }
     );
-    useRoleCapabilities.mockClear().mockReturnValue({
-      groupedRoleCapabilitiesByType,
-      initialRoleCapabilitiesSelectedMap: 1,
+    useUserCapabilities.mockClear().mockReturnValue({
+      groupedUserCapabilitiesByType: groupedRoleCapabilitiesByType,
+      initialUserCapabilitiesSelectedMap: 1,
       capabilitiesTotalCount: 2
     });
   });
@@ -96,7 +109,7 @@ describe('UsersCapabilitiesCompareItems', () => {
 
     renderComponent({ setRolesToCompare });
 
-    await userEvent.click(screen.getByText(users["11"].username))
+    await userEvent.click(screen.getByText(users[0].username))
 
     await userEvent.click(screen.getByText(selectedMemberOptions[0].label));
     await userEvent.click(screen.getByText(roles[0].name));