+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * [UICONSET-180](https://folio-org.atlassian.net/browse/UICONSET-180) Update network ordering permission set to include publications permissions.
 * [UICONSET-166](https://folio-org.atlassian.net/browse/UICONSET-166) Compare Authorization Role capabilities in full screen view.
 * [UICONSET-159](https://folio-org.atlassian.net/browse/UICONSET-159) Add Authorization Roles to the consortium manager.
+* [UICONSET-167](https://folio-org.atlassian.net/browse/UICONSET-167) Compare User capabilities in full screen view
 
 ## [1.1.0](https://github.com/folio-org/ui-consortia-settings/tree/v1.1.0) (2024-03-20)
 [Full Changelog](https://github.com/folio-org/ui-orders/compare/v1.0.2...v1.1.0)